pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "alleycat"
version = "0.6.0"
description = "A command line tool for chat conversations with LLMs"
authors = [
    {name = "Andrew", email = "andrew@groat.nz"},
]
dependencies = [
    "openai>=1.12.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "pydantic>=2.6.1",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.1",
    "platformdirs>=4.0.0",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing",
    "Topic :: Utilities",
]

[project.scripts]
alleycat = "alleycat_apps.cli.main:app"
alleycat-admin = "alleycat_apps.cli.admin_cmd:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.2.1",
    "mypy>=1.8.0",
    "types-PyYAML>=6.0.12.12",
    "build>=1.2.2.post1",
    "twine>=6.1.0",
]

[tool.setuptools]
include-package-data = true
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["alleycat_apps*", "alleycat_core*"]
namespaces = false

[tool.setuptools.package-data]
"*" = ["*.json", "*.yaml", "*.yml", "*.txt", "*.md"]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
addopts = "-ra -v"
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "requires_api: marks tests that make real API calls and should only run with --api flag",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "D",   # pydocstyle
]
ignore = ["D203", "D213", "E402"]

[dependency-groups]
dev = [
    "tomli>=2.2.1",
]

[project.urls]
"Homepage" = "https://github.com/avowkind/alleycat"
"Bug Tracker" = "https://github.com/avowkind/alleycat/issues"
"Documentation" = "https://github.com/avowkind/alleycat/tree/main/docs" 
